create RBAC for accessing claimed cluster
When a ClusterClaim is assigned a cluster, the clusterclaim
controller will create a Role and RoleBinding in the namespace
of the ClusterDeployment. The Role will grant full permissions for
Hive resources and GET permissions for the admin kubeconfig and
password secrets. The RoleBinding will assign the role to the
subjects specified in the ClusterClaim.

https://issues.redhat.com/browse/CO-1070

+++ pkg/resource/delete.go
@@ -7,12 +7,13 @@ import (
 	log "github.com/sirupsen/logrus"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	hivev1 "github.com/openshift/hive/pkg/apis/hive/v1"
 )
 
 // DeleteAnyExistingObject will look for any object that exists that matches the passed in 'obj' and will delete it if it exists
-func DeleteAnyExistingObject(c client.Client, key client.ObjectKey, obj runtime.Object, logger log.FieldLogger) error {
+func DeleteAnyExistingObject(c client.Client, key client.ObjectKey, obj hivev1.MetaRuntimeObject, logger log.FieldLogger) error {
 	logger = logger.WithField("object", key)
 	switch err := c.Get(context.Background(), key, obj); {
 	case apierrors.IsNotFound(err):
@@ -21,6 +22,10 @@ func DeleteAnyExistingObject(c client.Client, key client.ObjectKey, obj runtime.
 	case err != nil:
 		logger.WithError(err).Error("error getting object")
 		return errors.Wrap(err, "error getting object")
+	}
+	if obj.GetDeletionTimestamp() != nil {
+		logger.Debug("object has already been deleted")
+		return nil
 	}
 	logger.Info("deleting existing object")
 	if err := c.Delete(context.Background(), obj); err != nil {
